cleanup, docs
