+++ crates/turbopack-ecmascript/src/side_effect_optimization/locals/mod.rs
@@ -1,0 +1,3 @@
+pub mod chunk_item;
+pub mod module;
+pub mod reference;
