+++ crates/turbopack-ecmascript/src/side_effect_optimization/reexport_facade/mod.rs
@@ -1,0 +1,2 @@
+pub mod chunk_item;
+pub mod module;
